+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2
+
+- Added `columns` option to `each_row` method. Allows us to take advantage of the column projection feature of the parquet crate.
+- General refactoring to improve readability and maintainability.
+
 ## 0.0.1
 
 Initial release.